do_lzx_decompress: fix warnings

+++ src/decomp.cpp
@@ -169,9 +169,9 @@ string do_lzx_decompress(const string_view& compdata, uint32_t size) {
         uint32_t complen;
 
         if (i == 0)
-            complen = num_chunks > 1 ? offsets[0] : data.length();
+            complen = num_chunks > 1 ? offsets[0] : (uint32_t)data.length();
         else if (i == num_chunks - 1)
-            complen = data.length() - offsets[i - 1];
+            complen = (uint32_t)data.length() - offsets[i - 1];
         else
             complen = offsets[i] - offsets[i - 1];
